host abstraction: add host-config-allegro
If we need to get rid of of allegro4 library (like ESL wanted),
it is much better if we build an abstraction layer, independent
of any services provided by the library. Yes, it is a bit worse
from performance point of view, but with modern machines the
penalty is negligible.

I can see, that the emulator directly uses the following allegro
services:

- config file reading
- keyboard
- mouse
- joystick
- sound
- graphics

Let's start to build it from the config abstraction layer.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>