host-config: switch to libconfini
If we finally plan to implement LibSDL support (what ESL wanted),
then we need some support for the configuration file since libsdl
does not has it (vs liballegro, which has).

Introduce confini library. It is distributed under GPL-3, so
compatible.

The sources are taken from
https://github.com/madmurphy/libconfini.git. I did not take the
whole repo, just src/confini.{h,c} and added missed includes to
the .h. The commit id is
596723a39e223ec825069e91ae7c70f67fb5e281 (tag 1.4.2).

And rewrite host-config.c to use it.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>