main: refactor turbo mode logic, split to functions
The turbo mode and turbo boot logic was not very obvious for me,
so first take it out of the main loop and split into small
functions with the self-decribing names. The behaviour on states
like "entering turbo mode" or "exiting turbo mode" is more clear
now.

The initial state (needed because of the -T/-t command line
switches) is set one time in the beginning of the main loop.

As the side effect the patch

- fixes no-SOUND build (timer interrupts are checked now);
- conditional compilation is more clear as well;
- defines pr_* format;

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>