Makefile: make it possible to define compile options from cmdline
Some of the compile-time options were defined in the main include
file (korvet.h) so to change them it was necessary to edit it.

Make them as 'make' variables and set then CFLAGS (-D)
accordingly.

Because the defaults are different (DBG, SOUND, LAN_SUPPORT
enabled by default, but EGA, WAV and TRACETIMER disabled) you
should trigger the options differently -- to disable default
assign it 'no', to enable (disabled default) -- 'yes'.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>