main: switch to event based keyboard handling
It is still sort of polling, we do not sleep while waiting for an
event since not everything is yet event based, but the main
handling is in place.

The patch changes only the main loop handling, debugger and vm
scanning are still old (so it works incorrectly since the keys
from the debugger are feeded to the queue as well and processed
later).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>