keyboard: update VM keyboard state from the main event handler
This gets rid of one more place of using directly the allegro's
kbd array. This may change the behaviour since the scanning
timings are different, but hopefully it is ok here.

The patch reuses internal representation of the copy of the key
array, KeyAlias. And it required to change a bit alias
updating (defined by the AliasTab) since the state is tracked
now.

Only the keys not managed by the main event loop are forwarded to
the VM.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>