dbg: switch from readkey() to host event queue key handling
There are 6 places where keyboard was read with direct calls to
allegro4 library (readkey()).

Switch it to use the event with host_event_wait(), then filter
only keypresses. The code it transformed to the value expected
from allegro4 and other parts are untouched.

The HELP output chnaged in a bit more hakish way. Since it
invoked readkey() only for pause, it is splitted into 2 phases --
display the help on F1 and then switch to normal screen by any
other key. To avoid a lot of code changes, it uses a flag of help
display and it checked specially when a new key event is coming.

The event queue is flushed now (instead of pausing workaround)
before and after running dbg (to avoid sneaking extra keys
from/to VM).

simulate_keypress() is replaced directly with the
host_event_kbd_simulate() introduced before.

Hopefully, not a lot is broken, but since I'm not going to use
the debugger a lot, basically I do not care. Keep it just for the
sake of completeness.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>