host abstraction: add sound wrapper
To switch to other libraries we need to wrap them into our own
abstraction layer. Implement it.

allegro5 uses events mechanism to inform about the available
buffer (and has the old polling mechanism as well), while SDL
uses callback method. Since the buffer passed to callback is
valid only until callback exits, it is pretty clumsy to implement
asynchronious scheme with events based on that (it is also
possible, there will be a shared condition variable, the callback
would wait on it, and the event handler should call some
finalization function which signals on it).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>