host abstraction: implement and use host timers layer
This is a really thin layer on top on the allegro4 timer interface
since it is possible to have the same callback + context way there.

Originally I planned to use allegro5 approach, generating event, for
that, but a) the callback way looks cleaner, b) SDL uses callback
way, c) the existing code is callback-based so the change is really
trivial.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>