host events: implement joystick events
Add only one event for the joystick state. Since joystick requires
explicit polling on allegro4, start a thread for it.

The interface should be polished later, since now it requires
explicit initialization with joystick number parameter from the main
code. The function, which displays joystick information, is taken
from the original code almost as is, but printf changed to pr_info.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>