main joystick: switch to event model.
Hanlde joystick event in the main loop. The VMs joystick.c changed
accordingly, it requires just to get the update from the main loop,
save the state and generate port value when VM reads it.

As mentioned in the previoius patch, it requires explicit joystick
initialization and shutdown, which should be fixed by passing some
common state to the host_init().

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>