screen.c: refactor SCREEN_SetGraphics()
Make SCREEN_SetGraphics() to reuse common parts for different
mode settings. It a) calculates main window size in pixels; b)
calculates VM screen area size in pixels c) sets the video
mode. The sizes depend of mode and scaling factor, so organize
them separately. In order to do it, the patch gets rid of the
hardcoded scaling factor and refactors the scaling initialization
handling: decouple it from config file initialization. It used a
global variable as a scaling flag. Now use functions to handle it
-- on config parsing inform the main code of the scaling
information then pass the scaling factor to the screen init
function, so the implementation is hidden in the screen.c

The config file format kept compatible, but it's possible later
to change the semantics -- instead of on/off flag use it as a
factor to be able to do not only 1x/2x scaling but Nx.

Initialization steps (creating the bitmaps and setting color
depth) are moved to the init function for cleanup reasons.

Use defines for screen sizes (VM resolution, debug screen size).

The patch keeps the old behaviour -- does not update the screen
when in DBG mode. In this case if we enter debugger later after
scalling change the screen requires manual redraw (C-l), but it
is the same in the old code.

There is one external place, depending from the scaling
factor (WriteBMP), use a accessor function there as well.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>