host abstraction: initial implementation of host-graphics
Start the graphics abstraction. First step is to abstract bitmap
handling and VM screen -> real screen updates and mode setting.

The patch creates host_g_image independent bitmap abstraction
which is able to provide raw pixel line buffer pointer and blit
the bitmap to the physical screen.

The patch also abstracts set_gfx_mode() and sets the color depth
on init. The depth is hardcoded to 8.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>