screen: create vm_image abstraction and use abstract bitmap
Hide all the scaling related operation under internal vm_image
abstraction (represents virtual machine screen). The vm_image
implementation takes in use the abstract bitmap, introduced by
the previous patch.

The patch refactores SCREEN_ShowScreen() to handle every line
handling in the one function. There is some extra copying in the
implementation (the code first prepares the pixel buffer for the
line, then it gets copying with scaling to the target bitmap),
but it should not be a problem -- in the original code in case of
2x scaling, it always did the copying from the 1x bitmap. Now it
looks the same for both versions and handling bigger buffer on
copying must be more cache friendly.

Since now some structures are generated dynamically, we need to
have SCREEN_destroy() call to clean them up.

The original code missed bmp_unwrite_line(), now it calles it
with host_g_image_finish() (did not cause problems in the
original code, but should be there according to the allegro
doc).

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>