host graphics: implement and take in use palette wrapper
The palette handling pretty similar in allegro and SDL, so
implement a simple inline wrappers on top of the existing allegro
PALETTE type (which is an array of RGB structures). For SDL it
will become an array of SDL_Color structures which are a bit
different, they contain alpha channel. Implement init() call to
init the alpha to 0xFF in the future.

The wrapper uses vsync == true unconditionally (where it was not
used for the main palette updates, only for DBG), but it doesn't
look causing any problems.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>