screen/graphics: switch to abstract bmp dumping code
do not call directly allegro's save_bmp(), but wrap it with the
host abstraction.

This patch also makes some simplifications to the original
dumping code: there is no point to create temporary bitmap and
dump the real screen there, since our in-memory copy should
contain the actual version of it -- AllScreenUpdateFlag +
SCREEN_ShowScreen() should take care of it.

On the other hand, there is no point to use our copy of _subset_
of the palette, when we can get the real palette in the
abstraction layer. For SDL the palette will be part of the image
bitmap structure.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>