queue: new() API accepts actual number of available elements
Because of the internal implementation one element is used as
dummy to distinguish between empty and full cases. But it is not
obvious from the API (described in the comment).

It is more clear if the API actually accepts the number of
available elements for the user.

The patch also changes the client code to make the internal size
as power of two.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>