darray: add the data structure
The queue is not well suitable for further sound routines
improvements, since it requires random array access because some
frame boundaries may overlap in ticks->pcm conversion.

Implement a self increasing dynamic array which allows to add
data to it as to a buffer and read from random place, not only
from the end.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>