timer: fix DoTimer to process all the CPU time budget
CPU time processing loop in main takes into account, that the
consumed cycles may be not multiple of the clock
granularity (50Hz here, or 50000 cpu cycles per 1 emulator tick
for the 2.5Mhz CPU frequency -- 50000 * 50 == 2500000), it counts
the consumed cycles, then decreases the sum by 50000 leaving the
rest to the next iteration.

The timer code is driven by the CPU clock and has a similar
problem, just it consumes less. It has 2MHz frequency, or for
every 5 CPU cycles the timer makes 4 cycles. So the CPU time is
proceeded at 5 cycles boundaries and after handling it, some (not
multiple of 5) is left unprocessed till the next round.

The code just droped the unprocessed time.

So, let's accumulate it for futhre processing.

Also fix the DoTimer() signature. The return value does not make
sense at the moment.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>