queue: separate push/pop locking and export lock operations
Having one lock for both writing and reading (pushing and popping)
may cause lock contention in some cases. The advantage of the
underlying ring buffer is that reader accesses only the head of the
queue and writer only the tail, so they can be locked
separately (for the case when there are several readers or writers).

So, implement separate locking.

The patch also splits push/pop operations to normal (performs
locking) and locked versions to allow the client code to lock/unlock
the queue once before a batch of operations.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>