sound: redesign sound generator to fetch ticks from the callback
Make the sound generation driven not from main by the 50Hz clock,
but by the real time clock of the sound subsystem.

See the comment to the main function:

 * Converts timer ticks to PCM samples.
 *
 * The idea is pretty similar to the DoTimer() logic, but since
 * the frequencies are not devisible so nicely, fraction
 * arithmetic is used.
 *
 * It is driven by the sound clock, called from the sound
 * callback.
 *
 * Since the timer tick length is 500ns, sound sample is
 * ~22675.74ns, one sound sample contains ~90,7 ticks (where ~0.7
 * is 15500 / 22050, 15500 the reminder of timer frequency /
 * sound frequency, 2000000 % 22050 = 15500). Or, when even
 * number of ticks consumed, an error of ~0.7 is introduced. The
 * algorithm accumulates the error, and as soon as it is enough
 * for an even tick, consumes one extra tick and corrects the
 * error.
 *
 * Locking should not be needed, there is only one tick consumer.
 * But in possible case of underrun, if sound subsystem invokes
 * the callback, it looks better to wait the old one to finish.

So, there is no need to wait in the main anymore, removed. It is
other way around -- the sound callback waits for the ticks if there
are not enough of them to fill the buffer.

The tick code converted back to the queue, since the algorithm
consumes all the ticks sequentially. Looks like the buffer can grow
quite a lot in case of mute starting.

Removes also direct buffer manipulation from tooling, there is no
mirror buffer anymore and the sound muting method flushes the
buffers itself.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>